x86, boot: add mmio serial during compressed boot
Allows "console=uart[8250],mmio[32],0xADDR[,BAUDn8[,BASE_BAUD]]" to
be recognized during compressed boot early console setup, and during
boot console setup. Replaces defines with common serial defines. Adds
suport for mmio-based serial devices to compressed-boot early console,
and plumbs support for defining the base baud rate for UART clock
calculations (since mmio serial cards may not have the standard rate,
resulting in incorrect baud rates for mmio devices).

+++ arch/x86/boot/compressed/misc.h
@@ -53,7 +53,6 @@ int cmdline_find_option(const char *option, char *buffer, int bufsize);
 int cmdline_find_option_bool(const char *option);
 #endif
 
-
 #if CONFIG_RANDOMIZE_BASE
 /* aslr.c */
 unsigned char *choose_kernel_location(unsigned char *input,
@@ -75,10 +74,10 @@ unsigned char *choose_kernel_location(unsigned char *input,
 
 #ifdef CONFIG_EARLY_PRINTK
 /* early_serial_console.c */
-extern int early_serial_base;
+extern unsigned long early_serial_base;
 void console_init(void);
 #else
-static const int early_serial_base;
+static const unsigned long early_serial_base;
 static inline void console_init(void)
 { }
 #endif
